trace_test rename, import constants::*
