+++ sk-core/src/trace/tests/mod.rs
@@ -1,4 +1,4 @@
-mod exported_trace_test;
 mod pod_owners_map_test;
+mod trace_test;
 
 use sk_testutils::*;
